+++ app/ante/EthGasConsumeDecorator.go
@@ -1,0 +1,96 @@
+package ante
+
+import (
+	"github.com/ethereum/go-ethereum/common"
+	ethcore "github.com/ethereum/go-ethereum/core"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
+	sdkerrors "github.com/okex/exchain/libs/cosmos-sdk/types/errors"
+	"github.com/okex/exchain/libs/cosmos-sdk/x/auth"
+	"github.com/okex/exchain/libs/cosmos-sdk/x/auth/types"
+	evmtypes "github.com/okex/exchain/x/evm/types"
+	"math/big"
+)
+
+
+// EthGasConsumeDecorator validates enough intrinsic gas for the transaction and
+// gas consumption.
+type EthGasConsumeDecorator struct {
+	ak        auth.AccountKeeper
+	sk        types.SupplyKeeper
+	evmKeeper EVMKeeper
+}
+
+// NewEthGasConsumeDecorator creates a new EthGasConsumeDecorator
+func NewEthGasConsumeDecorator(ak auth.AccountKeeper, sk types.SupplyKeeper, ek EVMKeeper) EthGasConsumeDecorator {
+	return EthGasConsumeDecorator{
+		ak:        ak,
+		sk:        sk,
+		evmKeeper: ek,
+	}
+}
+
+// AnteHandle validates that the Ethereum tx message has enough to cover intrinsic gas
+// (during CheckTx only) and that the sender has enough balance to pay for the gas cost.
+//
+// Intrinsic gas for a transaction is the amount of gas
+// that the transaction uses before the transaction is executed. The gas is a
+// constant value of 21000 plus any cost inccured by additional bytes of data
+// supplied with the transaction.
+func (egcd EthGasConsumeDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
+	msgEthTx, ok := tx.(evmtypes.MsgEthereumTx)
+	if !ok {
+		return ctx, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "invalid transaction type: %T", tx)
+	}
+
+	// sender address should be in the tx cache from the previous AnteHandle call
+	address := msgEthTx.From()
+	if address.Empty() {
+		panic("sender address cannot be empty")
+	}
+
+	// fetch sender account from signature
+	senderAcc, err := auth.GetSignerAcc(ctx, egcd.ak, address)
+	if err != nil {
+		return ctx, err
+	}
+
+	if senderAcc == nil {
+		return ctx, sdkerrors.Wrapf(
+			sdkerrors.ErrUnknownAddress,
+			"sender account %s (%s) is nil", common.BytesToAddress(address.Bytes()), address,
+		)
+	}
+
+	gasLimit := msgEthTx.GetGas()
+	gas, err := ethcore.IntrinsicGas(msgEthTx.Data.Payload, []ethtypes.AccessTuple{}, msgEthTx.To() == nil, true, false)
+	if err != nil {
+		return ctx, sdkerrors.Wrap(err, "failed to compute intrinsic gas cost")
+	}
+
+	// intrinsic gas verification during CheckTx
+	if ctx.IsCheckTx() && gasLimit < gas {
+		return ctx, sdkerrors.Wrapf(sdkerrors.ErrOutOfGas, "intrinsic gas too low: %d < %d", gasLimit, gas)
+	}
+
+	// Charge sender for gas up to limit
+	if gasLimit != 0 {
+		// Cost calculates the fees paid to validators based on gas limit and price
+		cost := new(big.Int).Mul(msgEthTx.Data.Price, new(big.Int).SetUint64(gasLimit))
+
+		evmDenom := sdk.DefaultBondDenom
+
+		feeAmt := sdk.NewCoins(
+			sdk.NewCoin(evmDenom, sdk.NewDecFromBigIntWithPrec(cost, sdk.Precision)), // int2dec
+		)
+
+		err = auth.DeductFees(egcd.sk, ctx, senderAcc, feeAmt)
+		if err != nil {
+			return ctx, err
+		}
+	}
+
+	// Set gas meter after ante handler to ignore gaskv costs
+	newCtx = auth.SetGasMeter(simulate, ctx, gasLimit)
+	return next(newCtx, tx, simulate)
+}
